Maven化
Spigot1.8.2よりMaterialのLOCKED_CHESTが消えてる ( https://hub.spigotmc.org/jira/browse/SPIGOT-370 )

getTargetBlockが競合していたのでそれも修正
