+++ .gitignore
@@ -1,0 +1,5 @@
+bin
+setblock
+pluginsb
+pluginss
+/target/
