--- bukkit/.gitignore
@@ -1,4 +1,0 @@
-bin
-setblock
-pluginsb
-pluginss
